Don't inherit from Account

+++ contracts/Barrack.sol
@@ -9,7 +9,7 @@ contract Barrack is BuildingFactory, Soldier {
     using SafeMath for uint;
     using SafeMath for uint;
 
-    function createBarrack(uint16 _x, uint16 _y) public {
+    function createBarrack(uint _x, uint _y) public {
         _createBuilding(msg.sender, "Barrack", _x, _y);
         levelOfSoldier[msg.sender] = 1;
     }
@@ -50,82 +50,4 @@ contract Barrack is BuildingFactory, Soldier {
         }
     }
 
-    // // return 0 if failed (maybe already creating or not enough resource) otherwise return leveluptime
-    // function startLevelUpSoldier(string memory _name) public returns(uint) {
-    //     address _owner = msg.sender;
-    //     if (ownerStartLevelUpTime[_owner] != 0) return uint(0);// check if there is already leveling up soldiers
-    //     bool enoughResource;
-    //     uint levelOfSoldier;
-    //     uint numOfSoldier;
-    //     if (keccak256(bytes(_name)) == keccak256(bytes("Cavalry"))) {
-    //         enoughResource = _upgradeCavalry(_owner);
-    //         _name = "Cavalry";
-    //         levelOfSoldier = levelOfCavalry[_owner];
-    //         numOfSoldier = numOfCavalry[_owner];
-    //     }
-    //     // else if (keccak256(bytes(_name)) == keccak256(bytes("Pikemen"))) {
-    //     //     enoughResource = _upgradePikemen(_owner);
-    //     //     _name = "Pikemen";
-    //     //     levelOfSoldier = levelOfPikemen[_owner];
-    //     //     numOfSoldier = numOfPikemen[_owner];
-    //     // }
-    //     // else if (keccak256(bytes(_name)) == keccak256(bytes("Infantry"))) {
-    //     //     enoughResource = _upgradeInfantry(_owner);
-    //     //     _name = "Infantry";
-    //     //     levelOfSoldier = levelOfInfantry[_owner];
-    //     //     numOfSoldier = numOfInfantry[_owner];
-    //     // }
-    //     // else if (keccak256(bytes(_name)) == keccak256(bytes("Archer"))) {
-    //     //     enoughResource = _upgradeArcher(_owner);
-    //     //     _name = "Archer";
-    //     //     levelOfSoldier = levelOfArcher[_owner];
-    //     //     numOfSoldier = numOfArcher[_owner];
-    //     // }
-    //     if(enoughResource == false) return uint(0);
-    //     ownerStartLevelUpTime[_owner] = now;
-    //     return levelUpSoldierTime * numOfSoldier * uint(levelOfSoldier);
-    // }
-
-    // function updateLevelUpSoldier(string memory _name, address _owner) public returns(uint) {
-    //     if (ownerStartLevelUpTime[_owner] == 0) return 0;
-    //     uint levelOfSoldier;
-    //     uint numOfSoldier;
-    //     if (keccak256(bytes(_name)) == keccak256(bytes("Cavalry"))) {
-    //         levelOfSoldier = levelOfCavalry[_owner];
-    //         numOfSoldier = numOfCavalry[_owner];
-    //     }
-    //     // else if (keccak256(bytes(_name)) == keccak256(bytes("Pikemen"))) {
-    //     //     levelOfSoldier = levelOfPikemen[_owner];
-    //     //     numOfSoldier = numOfPikemen[_owner];
-    //     // }
-    //     // else if (keccak256(bytes(_name)) == keccak256(bytes("Infantry"))) {
-    //     //     levelOfSoldier = levelOfInfantry[_owner];
-    //     //     numOfSoldier = numOfInfantry[_owner];
-    //     // }
-    //     // else if (keccak256(bytes(_name)) == keccak256(bytes("Archer"))) {
-    //     //     levelOfSoldier = levelOfArcher[_owner];
-    //     //     numOfSoldier = numOfArcher[_owner];
-    //     // }
-    //     if (now >= ownerStartLevelUpTime[_owner].add(levelUpSoldierTime * numOfSoldier * uint(levelOfSoldier))) {
-    //         if (keccak256(bytes(_name)) == keccak256(bytes("Cavalry"))) {
-    //             levelOfCavalry[_owner]++;
-    //         }
-    //         else if (keccak256(bytes(_name)) == keccak256(bytes("Pikemen"))) {
-    //             levelOfPikemen[_owner]++;
-    //         }
-    //         else if (keccak256(bytes(_name)) == keccak256(bytes("Infantry"))) {
-    //             levelOfInfantry[_owner]++;
-    //         }
-    //         else if (keccak256(bytes(_name)) == keccak256(bytes("Archer"))) {
-    //             levelOfArcher[_owner]++;
-    //         }
-    //         ownerStartLevelUpTime[_owner] = 0;
-    //         _updatePower(_owner);
-    //         return 0;
-    //     }
-    //     else {
-    //         uint remainingTime = levelUpSoldierTime * numOfSoldier * uint(levelOfSoldier) + ownerStartLevelUpTime[_owner] - now;
-    //         return remainingTime;
-    //     }
-    // }
 }
